+++ R/aaa.R
@@ -1,0 +1,3 @@
+#' @useDynLib isocubes, .registration=TRUE
+NULL
+