+++ src/hash-color.h
@@ -1,0 +1,2 @@
+int hash_color(const unsigned char *key);
+